server/requirements.txt
# Python dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.116.0
uvicorn>=0.35.0
pymongo>=4.14.0
python-dotenv>=1.1.0
slack-sdk>=3.36.0
flask>=3.0.0
flask-cors>=4.0.0
aiohttp>=3.8.0
requests>=2.31.0
